promoted all but __call__ to message
